move redirect logic to docs route
