+++ app/routes/application.js
@@ -6,14 +6,8 @@ import localStorageProxy from 'percy-web/lib/localstorage';
 import {DO_NOT_FORWARD_REDIRECT_ROUTES} from 'percy-web/router';
 import EnsureStatefulLogin from 'percy-web/mixins/ensure-stateful-login';
 import isDevWithProductionAPI from 'percy-web/lib/dev-auth';
-import utils from 'percy-web/lib/utils';
-import {Promise} from 'rsvp';
 
 export const AUTH_REDIRECT_LOCALSTORAGE_KEY = 'percyAttemptedTransition';
-
-const REDIRECTS = {
-  '/docs/clients/javascript/react-storybook': '/docs/storybook-for-react',
-};
 
 export default Route.extend(ApplicationRouteMixin, EnsureStatefulLogin, {
   session: service(),
@@ -21,31 +15,8 @@ export default Route.extend(ApplicationRouteMixin, EnsureStatefulLogin, {
   raven: service(),
   currentUser: alias('session.currentUser'),
 
-  _decideDocsRedirect(transition) {
-    const targetPage = transition.intent.url;
-    if (targetPage in REDIRECTS) {
-      return utils.setWindowLocation(`https://docs.percy.io${REDIRECTS[targetPage]}`);
-    } else {
-      return utils.setWindowLocation('https://docs.percy.io');
-    }
-  },
-
   beforeModel(transition) {
     this._super(...arguments);
-
-    // If the target route contains anything related to the old docs, don't continue.
-    // This returns a promise because beforeModel blocks on a returned promise.
-    // If a promise is not returned, it starts the transition but continues loading the app
-    // causing visual jank.
-    // The promise does not resolve, as we do not want the app to proceed further
-    // if we are redirecting
-    if (window.location.origin.includes('percy') && transition.intent.url.includes('docs')) {
-      return new Promise(() => {
-        this._decideDocsRedirect(transition);
-        // window.location.replace('https://docs.percy.io');
-      });
-    }
-
     if (!this.get('session.isAuthenticated')) {
       this._storeTargetTransition(transition);
